+++ src/TagPlugin.groovy
@@ -4,6 +4,7 @@ class TagPlugin implements TerraformPlanCommandPlugin,
 
     private static variableName
     private static disableOnApply = false
+    private static writeToFile = false
     private static List tagClosures = []
 
     public static init() {
@@ -31,7 +32,11 @@ class TagPlugin implements TerraformPlanCommandPlugin,
         String variableName = getVariableName()
         Map tags = getTags(command)
 
-        command.withVariable(variableName, tags)
+        if (writeToFile) {
+            command.withVariableFile(variableName, tags)
+        } else {
+            command.withVariable(variableName, tags)
+        }
     }
 
     public static withVariableName(String variableName) {
@@ -79,6 +84,11 @@ class TagPlugin implements TerraformPlanCommandPlugin,
         return this
     }
 
+    public static writeToFile() {
+        writeToFile = true
+        return this
+    }
+
     private static getVariableName() {
         return variableName ?: 'tags'
     }
@@ -93,5 +103,6 @@ class TagPlugin implements TerraformPlanCommandPlugin,
         tagClosures = []
         variableName = null
         disableOnApply = false
+        writeToFile = false
     }
 }